Insurance And Policie Working.

+++ src/main/java/com/technical/assessment/security/service/UserPrinciple.java
@@ -16,7 +16,7 @@ public class UserPrinciple implements UserDetails {
 
     private static final long serialVersionUID = 1L;
 
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -29,7 +29,7 @@ public class UserPrinciple implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserPrinciple(int id, String name,
+    public UserPrinciple(Long id, String name,
                          String username, String email, String password,
                          Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
@@ -55,7 +55,7 @@ public class UserPrinciple implements UserDetails {
         );
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
